requirements.txt
# Pin these dependencies
Django==1.11.21
requests==2.13.0
requests-oauthlib==0.7.0
thrift==0.10.0
thrift_connector==0.24
djangorestframework==3.8.2
django-webpack-loader==0.6.0

wagtail==2.0
wagtailfontawesome==1.1.3

jupyter==1.0.0
papermill==1.0.1

-e "."
